server_api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.9
sqlalchemy==2.0.43
pydantic==2.10.1
pydantic-settings==2.10.1
psycopg[binary]==3.2.9
numpy==1.26.4
opencv-python-headless==4.10.0.84
onnxruntime==1.19.2
insightface==0.7.3
reportlab==4.2.2
python-dotenv==1.0.1
openai==1.68.2
google-generativeai==0.7.2
duckduckgo-search==7.5.5
langchain==0.3.27
langchain-core==0.3.83
langchain-openai==0.3.11
langchain-google-genai==2.0.0
langchain-community==0.3.21
langchain-text-splitters==0.3.11
faiss-cpu==1.8.0
pypdf==5.4.0
pytesseract==0.3.13
ultralytics==8.3.60
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 
